NoListView, DarkMode and iPad Mode

+++ Todo/Views/ListView.swift
@@ -12,17 +12,23 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item).onTapGesture {
-                    withAnimation(.linear){
-                        listViewModel.updateItem(item: item)
+        ZStack{
+            if listViewModel.items.isEmpty {
+                NoItemsView().transition(AnyTransition.opacity.animation(.easeInOut))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item).onTapGesture {
+                            withAnimation(.linear){
+                                listViewModel.updateItem(item: item)
+                            }
+                        }
                     }
-                }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }.listStyle(.automatic)
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-        }.listStyle(.automatic)
+        }
         .navigationTitle("Todo List 📝")
             .navigationBarItems(
                 leading: EditButton(),
